Move I/O routines for the PixelData class.
This creates `read()` and `write()` methods for the PixelData
class that can work with both WCS and Healpix, as well as
HDF5 and FITS formats.  These use the existence of the
PixelDistribution.wcs member to distinguish between WCS and
Healpix data, and they use the filename extension to determine
the format.

The serial I/O routines can now claim the shorter names
(read_wcs, write_healpix, etc) without conflicts.  The low-level
helper functions for gathering / broadcasting map chunks are
left in their current place outside the PixelData class, since
that class is becoming large.
