+++ src/toast/tests/ops_demodulate.py
@@ -11,8 +11,6 @@ from astropy import units as u
 from .. import ops as ops
 from .. import qarray as qa
 from ..observation import default_values as defaults
-from ..pixels import PixelData
-from ..pixels_io_healpix import write_healpix_fits
 from ..vis import set_matplotlib_backend
 from .helpers import (
     close_data,
