+++ src/toast/tests/ops_noise_estim.py
@@ -14,8 +14,6 @@ from .. import ops as ops
 from .. import qarray as qa
 from ..noise import Noise
 from ..observation import default_values as defaults
-from ..pixels import PixelData, PixelDistribution
-from ..pixels_io_healpix import write_healpix_fits
 from ..vis import set_matplotlib_backend
 from .helpers import (
     close_data,
