+++ src/toast/tests/ops_pixels_healpix.py
@@ -17,8 +17,6 @@ from ..intervals import IntervalList
 from ..mpi import MPI
 from ..noise import Noise
 from ..observation import default_values as defaults
-from ..pixels import PixelData, PixelDistribution
-from ..pixels_io_healpix import write_healpix_fits
 from ..vis import set_matplotlib_backend
 from .helpers import create_outdir
 from .mpi import MPITestCase
