+++ src/toast/tests/ops_sim_cosmic_rays.py
@@ -9,8 +9,6 @@ import numpy as np
 from .. import ops as ops
 from .. import rng
 from ..covariance import covariance_apply
-from ..pixels import PixelData, PixelDistribution
-from ..pixels_io_healpix import write_healpix_fits
 from ..vis import set_matplotlib_backend
 from .helpers import (
     close_data,
