+++ src/toast/tests/ops_sim_tod_atm.py
@@ -13,7 +13,6 @@ from .. import qarray as qa
 from ..atm import available_atm, available_utils
 from ..dipole import dipole
 from ..observation import default_values as defaults
-from ..pixels_io_healpix import write_healpix_fits
 from ..vis import set_matplotlib_backend
 from .helpers import close_data, create_ground_data, create_outdir
 from .mpi import MPITestCase
