+++ src/toast/tests/ops_sim_tod_dipole.py
@@ -11,7 +11,6 @@ from astropy import units as u
 from .. import ops as ops
 from .. import qarray as qa
 from ..dipole import dipole
-from ..pixels_io_healpix import write_healpix_fits
 from ..vis import set_matplotlib_backend
 from .helpers import close_data, create_healpix_ring_satellite, create_outdir
 from .mpi import MPITestCase
@@ -134,8 +133,8 @@ class SimDipoleTest(MPITestCase):
 
         toast_hit_path = os.path.join(self.outdir, "toast_hits.fits")
         toast_bin_path = os.path.join(self.outdir, "toast_bin.fits")
-        write_healpix_fits(data[binner.binned], toast_bin_path, nest=False)
-        write_healpix_fits(data[cov_and_hits.hits], toast_hit_path, nest=False)
+        data[binner.binned].write(toast_bin_path, healpix_nest=False)
+        data[cov_and_hits.hits].write(toast_hit_path, healpix_nest=False)
 
         rank = 0
         if self.comm is not None:
