+++ src/toast/tests/pixels_wcs.py
@@ -104,9 +104,9 @@ class PixelTest(MPITestCase):
                     self.outdir,
                     "data_sub{}_type-{}.{}".format(nsb, np.dtype(tp).char, suffix),
                 )
-                io.write_wcs(pdata, wcsfile)
+                pdata.write(wcsfile)
                 check = self._make_pixdata(dist, tp, n_value, zero=True)
-                io.read_wcs(check, wcsfile)
+                check.read(wcsfile)
                 nt.assert_equal(pdata.data, check.data)
 
                 if not available_pixell:
